pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eservice-ai-platform"
version = "1.0.0"
description = "eService AI Platform - Intelligent Customer Service System with LLM/RAG Integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "eService Team", email = "contact@eservice.com"}
]
keywords = ["fastapi", "microservices", "llm", "rag", "agent", "customer-service"]

dependencies = [
    # Core Framework
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    
    # Asyncio & Concurrency
    "httpx==0.25.2",
    "aioredis==2.0.1",
    "asyncpg==0.29.0",
    
    # Database
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.1",
    
    # Message Queue
    "aio-pika==9.1.3",
    "celery==5.3.4",
    
    # Caching & Redis
    "redis==5.0.1",
    
    # LLM Integration
    "openai==1.3.9",
    "langchain==0.1.6",
    "langchain-openai==0.0.7",
    "langchain-community==0.0.20",
    "langgraph==0.0.39",
    
    # RAG & Vector DB
    "pinecone-client==2.2.4",
    "milvus==2.3.5",
    "langchain-text-splitters==0.0.1",
    
    # Embeddings
    "sentence-transformers==2.2.2",
    
    # Monitoring & Observability
    "prometheus-client==0.19.0",
    "opentelemetry-api==1.21.0",
    "opentelemetry-sdk==1.21.0",
    "opentelemetry-exporter-prometheus==0.42b0",
    "opentelemetry-exporter-jaeger==1.21.0",
    "python-json-logger==2.0.7",
    
    # Data Validation & Serialization
    "python-multipart==0.0.6",
    "orjson==3.9.10",
    
    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "httpx==0.25.2",
    
    # Code Quality
    "black==23.12.0",
    "ruff==0.1.8",
    "mypy==1.7.1",
    
    # Utilities
    "python-dotenv==1.0.0",
    "tenacity==8.2.3",
    "structlog==23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest-env==1.1.3",
    "pytest-mock==3.12.0",
    "faker==21.0.0",
    "ipython==8.18.1",
    "ipdb==0.13.13",
]

docs = [
    "mkdocs==1.5.3",
    "mkdocs-material==9.5.3",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "W", "I", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
